final update souhe

+++ timesheet-ci/src/main/java/tn/esprit/spring/repository/EmployeRepository.java
@@ -7,6 +7,6 @@ import tn.esprit.spring.entities.Employe;
 
 
 @Repository
-public interface EmployeRepository extends CrudRepository<Employe, Integer> {
+public interface EmployeRepository extends CrudRepository<Employe, Long> {
 
 }
